--- NineYi.BurgerShop/Breads/WheatBread.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NineYi.BurgerShop.Breads
-{
-    /// <summary>
-    /// 全麥麵包
-    /// </summary>
-    public class WheatBread : Bread
-    {
-        public WheatBread()
-        {
-            this.Name = "Wheat Bread";
-        }
-    }
-}
