--- NineYi.BurgerShop/Breads/WhiteBread.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NineYi.BurgerShop.Breads
-{
-    /// <summary>
-    /// 白麵包
-    /// </summary>
-    public class WhiteBread : Bread
-    {
-        public WhiteBread()
-        {
-            this.Name = "White Bread";
-        }
-    }
-}
