+++ NineYi.BurgerShop/Burgers/Burger.cs
@@ -1,4 +1,4 @@
-﻿using NineYi.BurgerShop.Breads;
+﻿using NineYi.BurgerShop.Models.Breads;
 using NineYi.BurgerShop.Meats;
 using NineYi.BurgerShop.Veggies;
 using System;
