+++ NineYi.BurgerShop/Models/Breads/WhiteBread.cs
@@ -1,0 +1,17 @@
+﻿namespace NineYi.BurgerShop.Models.Breads
+{
+    /// <summary>
+    /// 白麵包。
+    /// </summary>
+    /// <seealso cref="NineYi.BurgerShop.Models.Breads.Bread" />
+    public class WhiteBread : Bread
+    {
+        /// <summary>
+        /// 初始化 <see cref="WhiteBread"/> 類別新的執行個體。
+        /// </summary>
+        public WhiteBread()
+        {
+            this.Name = "White Bread";
+        }
+    }
+}
